Add files via upload
Sure, here is a 3-sentence description of the Titanic classification project using Python through Colab:

Titanic Classification Project Using Python Through Colab

This project uses Python through Colab to build and evaluate a machine-learning model to predict passenger survival on the Titanic. The model is trained on a dataset of historical passenger data, including features such as age, sex, and class. The model is then evaluated on its ability to predict the survival of passengers on a held-out test set.
This project is a great way to learn about machine learning and Python, and it also provides a valuable contribution to the field of data science.